Polish cluster starting
